[ELK] Move functions from Enrich to Identities
This commit moves identity related functions from
Enrich class to Identities class to improve cohesion
and reduce LOC in Enrich class.
