+++ tests/tests.conf
@@ -1,2 +1,2 @@
 [ElasticSearch]
-url=http://localhost:9200
+url=https://admin:admin@localhost:9200
